backend/pyproject.toml
[project]
name = "atlas-backend"
version = "0.1.0"
description = "Atlas - Local RAG Knowledge Base Assistant Backend"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "ollama>=0.4.0",
    "chromadb>=0.4.22",
    "python-multipart>=0.0.6",
    "sqlalchemy>=2.0.0",
    "aiosqlite>=0.19.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.23.0",
    "python-docx>=1.1.0",
    "chardet>=5.2.0",
    "langchain-text-splitters>=0.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.26.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
